packages/grid/apps/network/pyproject.toml
[tool.poetry]
name = "PyGrid Domain"
version = "0.3.0"
description = ""
authors = ["Ionesio Junior <ionesiojr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gevent = "^20.6.2"
gevent-websocket = "^0.10.1"
gunicorn = "^20.0.4"
Flask = "^1.1.1"
Flask-Sockets = "^0.2.1"
numpy = "^1.18.5"
sqlitedict = "^1.6.0"
sqlalchemy_utils = "^0.36.8"
flask_sqlalchemy = "^2.4.4"
flask_migrate = "^2.5.3"
sqlalchemy_mixins = "^1.2.1"
PyJWT = "^1.7.1"
bcrypt = "^3.2.0"
loguru = "^0.5.3"
Flask-Executor = "^0.9.4"
Flask-Cors = "^3.0.10"
syft = { git = "https://github.com/OpenMined/PySyft", rev = "0.5.0rc1" }
terrascript = "^0.9.0"
PyInquirer = "^1.0.3"
boto3 = "^1.14.51"
textwrap3 = "^0.9.2"
requests-toolbelt = "0.9.1"
scipy = "^1.6.1"
tenseal = "^0.3.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = "^4.5.3"
pre-commit = "^2.6.0"
pytest = "^5.4.3"
docformatter = "^1.3.1"
flake8 = "^3.8.3"
flake8-comprehensions = "^3.2.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
